Stack smashing protection
Initialise the __stack_chk_guard value on x86_64 and aarch64 from
_start(). Given that we don't have a source of randomness, derive a weak
pseudo-random value using CPU ticks. This is a reasonable trade-off for
now without diving into the extra complexity and ABI changes getting
randomness from the host would entail.

On x86_64, additionally, -mstack-protector-guard=global is used to
instruct the compiler to use a plain global rather than a thread-local
variable. This means we do not need any TLS setup code at this stage.

__stack_chk_fail() is just a straight immediate abort, this could
possibly be improved in the future to show at least the caller address
if that is possible safely.

A test_ssp has been added which smashes the stack intentionally, thus
verifying that SSP is working as expected.