test_fpu: Fix stack corruption
c[] is only 64 bits (at sizeof (float) == 4), so the x86_64 variant was
happily loading and storing double the bits, thus silently corrupting
the stack (caught by SSP!). Extend to a 4-element vector, and test that
the vector multiplication operated on all 4 elements.

For aarch64 the code was fine, but extend it anyway to use a 4-element
vector so that we can compare against the same results.