Remove Genode bindings
The Genode bindings have been unsupported for some time, and require
extra work to be forward-ported to the upcoming support for multiple
toolchains. Remove them for now, they can be reinstated later if there
is interest.

Note that GENODE_ABI_TARGET is not removed, since this has been
allocated the value should not be re-used by any future bindings.